requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
pydantic[email]==2.9.0
email-validator==2.1.0
sqlalchemy==2.0.23
asyncpg==0.29.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.6
alembic==1.13.1
psutil==5.9.8
requests==2.31.0
